requirements.txt
apify-client==1.3.1
beautifulsoup4==4.12.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
ollama==0.1.5